Move changelog funcs to the release utils
